feat: add translation management for installation and selection
Add UI for translation managing. The available translations for downloading in the download page, and the installed translations in the installed page.

Adjust a little IO behavior to make things work correct

+++ src/utils/io.ts
@@ -119,7 +119,10 @@ async function resolveTranslation(extracted: string) {
 
   console.log(`Resolving translation ${translationId} at ${extracted} to ${translationFolder}`)
 
-  await fs.remove(translationFolder, { recursive: true })
+  if (await fs.exists(translationFolder)) {
+    await fs.remove(translationFolder, { recursive: true })
+  }
+
   await fs.mkdir(translationFolder, { recursive: true })
 
   const resolves: Promise<unknown>[] = []
@@ -163,7 +166,7 @@ async function installTranslation(translationId: string) {
 
 async function getInstalledTranslations() {
   if (!(await fs.exists(translationsPath))) {
-    return null
+    return undefined
   }
 
   const dirs = await fs.readDir(translationsPath)
